--- src/main/java/xyz/stasiak/cobudget/expense/Expense.java
@@ -1,9 +1,0 @@
-package xyz.stasiak.cobudget.expense;
-
-import org.springframework.data.annotation.Id;
-
-import java.time.LocalDate;
-
-record Expense(@Id Long id, int amount, LocalDate date, String category, String subcategory) {
-
-}
